fix rounding error for long sequences on one line
git-svn-id: svn+ssh://svn.internal.sanger.ac.uk/repos/svn/pathsoft/artemis/trunk@15133 ee4ac58c-ac51-4696-9907-e4b3aa274f04

+++ uk/ac/sanger/artemis/io/StreamSequence.java
@@ -296,7 +296,7 @@ public abstract class StreamSequence
   protected void appendChar(final char dna[])
   {
     int newlength = sequence_length + dna.length;
-    int numBytes  = Math.round( newlength/2.f );
+    int numBytes  = (int)Math.round( (double)newlength/2.d );
     if(numBytes > capacity())
       expandCapacity(numBytes);
